api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pykrx==1.0.45
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
supabase==2.7.2
feedparser==6.0.11
